Add errors necessary for RFML
